added attribution.py for importing functions
