+++ run_attribution.py
@@ -13,15 +13,13 @@ from sklearn.model_selection import train_test_split
 
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
 from torch.utils.data import DataLoader
 
 import src.models
+from src.models.attribution import to_np, compute_attr_x
 from src.dataset import SimpleDataset
 from src.preprocess import preprocess
 
-from captum.attr import Saliency, IntegratedGradients, DeepLiftShap, Lime, LayerGradCam, GuidedGradCam, InputXGradient, KernelShap, FeatureAblation, LayerAttribution, LRP, DeepLift
-from captum.attr import visualization as viz
 
 parser = argparse.ArgumentParser(description='Attribution ECG')
 parser.add_argument('--dataset_path', default='dataset/12000_btype_new.pkl', type=str, help='path to dataset')
@@ -59,51 +57,6 @@ def main():
     ## Compute attribution for an input
     attr_x = compute_attr_x(model, args.attr_method, input_tensor, args.n_samples)
     torch.save(attr_x, args.results_path + args.attr_method + "_" + str(args.input_idx) + ".pt")
-
-def compute_attr_x(model, attr_method, input_tensor, n_samples=200, absolute=False):
-    ## Select target (predicted label)
-    yhat = model(input_tensor)
-    softmax_yhat = F.softmax(yhat, dim=1)
-    prediction_score, pred_label_idx = torch.topk(softmax_yhat, 1)
-
-    attribution_dict = {"saliency": Saliency(model),
-                        "integrated_gradients": IntegratedGradients(model),
-                        "input_gradient": InputXGradient(model),
-                        "lrp": LRP(model),
-                        "lime": Lime(model),
-                        "kernel_shap": KernelShap(model),
-                        "deep_lift": DeepLift(model),
-                        "deep_lift_shap": DeepLiftShap(model),
-                        "gradcam": LayerGradCam(model, model.layer4),
-                        "guided_gradcam": GuidedGradCam(model, model.layer4),
-                        "feature_ablation": FeatureAblation(model),
-                        }
-
-    ## Load attribution function
-    attr_func = attribution_dict[attr_method]
-    
-    ## Conduct attribution method
-    if attr_method in ["lime", "kernel_shap"]:
-        attr_x = attr_func.attribute(input_tensor, target=pred_label_idx, n_samples=n_samples)
-    elif attr_method == "deep_lift_shap":
-        attr_x = attr_func.attribute(input_tensor, target=pred_label_idx, baselines=torch.randn([n_samples] + list(input_tensor.shape[1:])).cuda())
-    else:
-        attr_x = attr_func.attribute(input_tensor, target=pred_label_idx)
-
-    ## Interpolation for GradCAM
-    if attr_method == "gradcam":
-        attr_x = LayerAttribution.interpolate(attr_x, (1, input_tensor.shape[-1]))
-    
-    ## Absolute values of attribution
-    if absolute:
-        attr_x = torch.abs(attr_x)
-
-    return attr_x
-
-
-def to_np(x):
-    return x.detach().cpu().numpy()
-
 
 def setup(args):
     setup_gpu(args)
